requirements.txt
BeautifulSoup4
python-dotenv
python-telegram-bot
selenium